CR1: added pytest-cov module to requirements; renamed testing folder
